feature: component and vuex logic for adding codes
Created new store module for adding codes. So far request all possible
countries and get all cities based on a country. Created basic vue
component for adding codes, in the end it will be a dependent
dropdown, so based on a selected country a city can be selected from
a new list.