fix: Assign data-pos with watch now
Data-pos formerly were asigned in mounted hook, plain with no method
invokation. Problem was that when clicking the back button in the
browser, the mounted hook was not run / run too early, so that data-pos
assignment did not happen and carousel cards' css was not applied
(due to lack of data-pos selectors). Now a watcher watches the
unevenCodes computed prop and then runs the assignment code which
formerly was in mounted as a function - addDataPos method.