refactor: Use global mutations for codes modules
Added two mutations, addDataPositions and prettyCode, to the root store
and removed them from the modules (i.e. codes, continents, countries)
respectively. This change lead to creating global state.codes and codes
getter, and removing these props / objects from the modules
respectively.
Additionaly in continents and countries module it is checked wether
new_cache_key is set; if so the global setURLMessage mutation is
committed with a slugs bject as payload. In root store a getter will
set an error message that should be displayed to the user when an
invalid URL was entered (new_cache_key's value is only set when the URL
was invalid).
Due to a structural change in the res.data object adjusted response
handling in modules' action: Since new_cache_key was added in the API,
codes now need to be retrieved from res.data.data (formerly res.data).
To make the actions more readable, in the first .then addDataPositions
is committed and in another .then the prettyCode and setRecentStatus
mutations.