feature: Add async Comp to show invalid slugs
When a user entered an invalid slug for either continent or country or both, not only the next valig slug part is being requested (django), but from now on it is displayed to the user what they entered and what
instead is being shown to them. Only basic implementation, since styling still needs to be done.