refactor: Use HTML form tag for AddCode form
Changed div to form tag. No prevented form action at top. Since there
are two buttons in that form, the one which should NOT submit but
only cancel the form got tyoe button.
Now the cancelForm method don't refreshes the page but empties the input
fields instead.
Changed name of method which submits the form to be same as Login etc.
Minor clean-up of unnecessary commentary.