refactor: Reusable Auth form component
Used an object passed to AuthForm with form groups set to true or false
depending on wether they should be shown in the current view or not.
Ended up making only a reusable form component for anything form-auth-
related. It already feels a bit over-engineered and there's a possibilty
I take everything back.
But for now it works like this: In a view (i.e. has its own URL) several
props are passed to AuthForm, e.g. the heading or an object with all
possible form groups (input fields) as keys and Booleans as values, set
to true whenever the input field should be shown in a view. The action
button in AuthForm emits a custom event and passed the form data to the
parent. Any parent view then has its own function to validate and handle
the data (dispatches on actions to API).