refactor: Use Modal as component instead of views
Reasons for switching:
1. I don't need a separate URL for the the qrcode modal - I pass the data
to the modal and do not retrieve it inside of it (no request to show
single code resource). Therefor no unique URL is needed.
2. I still want to be able to use the browser's back button to close
the modal and not navigate back. I achieved this with beforeRouteLeave
guard - caveat: it won't work if the route that is navigated back to
is an external route (e.g. on first page visit). Maybe I will try to
implement it with beforeunload event listener, but for now it is fine
like this.
3. When opening a modal while it was a view in the background the codes
table was visible (which was fine) and a request to the code endpoint
was dispatched again which resulted in a distraction of the background
being loaded (with skeleton showing etc.). I tried to bypass that by
wrapping the Table component in component tags inside a keep-alive.
The problem with that is that the request is not being dispatched from
the Table comp but from the route component, which I would like to keep
like that for now.